Original upload
Here is the chess program code. I have not tested it for portability. The AI currently moves randomly. There may be some bugs I am unaware of but this code should be stable.

This was originally a programming assignment for a Computer Science class at BYU (CS 240 Advanced Programming Concepts).